requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
python-dotenv
setuptools
pymongo
pymongo[srv]==3.11
certifi
# -e .
